Move the filter tests to the Courses Page instead of the Course Instance table
The filtering is happening on the Courses Page level instead now.
