+++ src/client/components/pages/Courses/__tests__/CourseInstanceTable.test.tsx
@@ -6,19 +6,18 @@ import {
   AllByRole,
   getRoles,
   within,
-  fireEvent,
 } from 'test-utils';
 import { spy, SinonSpy } from 'sinon';
 import { cs50CourseInstance, es095CourseInstance } from 'testData';
-import { COURSE_TABLE_COLUMN, isSEASEnumToString, IS_SEAS } from 'common/constants';
-import OFFERED, { offeredEnumToString } from 'common/constants/offered';
+import { COURSE_TABLE_COLUMN } from 'common/constants';
 import CourseInstanceTable from '../CourseInstanceTable';
 import { tableFields } from '../tableFields';
-import * as filters from '../../Filter';
+import { FilterState } from '../filters.d';
 
 describe('CourseInstanceTable', function () {
   let updateSpy: SinonSpy;
-  let filterSpy: SinonSpy;
+  let openMeetingModalSpy: SinonSpy;
+  let openInstructorModalSpy: SinonSpy;
   const academicYear = 2020;
   const courseList = [
     cs50CourseInstance,
@@ -26,7 +25,6 @@ describe('CourseInstanceTable', function () {
   ];
   beforeEach(function () {
     updateSpy = spy();
-    filterSpy = spy(filters, 'listFilter');
   });
   describe('Header rows', function () {
     context('With all fields visible', function () {
@@ -46,7 +44,7 @@ describe('CourseInstanceTable', function () {
       ];
       let getAllByRole: BoundFunction<AllByRole>;
       const areaFilterLabel = 'The table will be filtered as selected in this area dropdown filter';
-      const isSEASFilterLabel = 'The table will be filtered as selected in this Is SEAS dropdown filter';
+      const isSEASFilterLabel = 'The table will be filtered as selected in this isSEAS dropdown filter';
       const fallOfferedFilterLabel = 'The table will be filtered as selected in this fall offered dropdown filter';
       const springOfferedFilterLabel = 'The table will be filtered as selected in this spring offered dropdown filter';
       beforeEach(function () {
@@ -115,49 +113,6 @@ describe('CourseInstanceTable', function () {
         strictEqual(isSEASFilter.length, 1, 'Error with isSEAS filter rendering');
         strictEqual(fallOfferedFilter.length, 1, 'Error with fall offered filter rendering');
         strictEqual(springOfferedFilter.length, 1, 'Error with spring offered filter rendering');
-      });
-      context('when the area dropdown filter is changed', function () {
-        it('calls the listFilter function once for each filter', function () {
-          const [, , thirdRow] = getAllByRole('row');
-          const utils = within(thirdRow);
-          const area = utils.getByLabelText(areaFilterLabel);
-          filterSpy.resetHistory();
-          fireEvent.change(area, { target: { value: 'AM' } });
-          strictEqual(filterSpy.callCount, 1);
-        });
-      });
-      context('when the isSEAS dropdown filter is called', function () {
-        it('calls the listFilter function once for each filter', function () {
-          const [, , thirdRow] = getAllByRole('row');
-          const utils = within(thirdRow);
-          const isSEAS = utils.getByLabelText(isSEASFilterLabel);
-          filterSpy.resetHistory();
-          fireEvent.change(isSEAS,
-            { target: { value: isSEASEnumToString(IS_SEAS.N) } });
-          strictEqual(filterSpy.callCount, 1);
-        });
-      });
-      context('when the fall offered dropdown filter is called', function () {
-        it('calls the listFilter function once for each filter', function () {
-          const [, , thirdRow] = getAllByRole('row');
-          const utils = within(thirdRow);
-          const fallOffered = utils.getByLabelText(fallOfferedFilterLabel);
-          filterSpy.resetHistory();
-          fireEvent.change(fallOffered,
-            { target: { value: offeredEnumToString(OFFERED.Y) } });
-          strictEqual(filterSpy.callCount, 1);
-        });
-      });
-      context('when the spring offered dropdown filter is called', function () {
-        it('calls the listFilter function once for each filter', function () {
-          const [, , thirdRow] = getAllByRole('row');
-          const utils = within(thirdRow);
-          const springOffered = utils.getByLabelText(springOfferedFilterLabel);
-          filterSpy.resetHistory();
-          fireEvent.change(springOffered,
-            { target: { value: offeredEnumToString(OFFERED.N) } });
-          strictEqual(filterSpy.callCount, 1);
-        });
       });
     });
     context('With no semester fields visible', function () {
